Resolve Code Review comments
This commit makes changes in line with most of Hunter's comments on the
repo's PR #1 Specifically, it adds a before_action filter and a private
`#set_user` method and remove manual actions of setting the user. It
changes the rides index view to swap <ul> tags for <p> tags. It deletes
this branch's empty UserController file.

The one unaddressed comment was regarding Rails default behavior with
request format specification and it needs the respond_to block to work.